+++ Sodium/GenericHash.swift
@@ -28,9 +28,9 @@ public class GenericHash {
         }
         var ret: CInt;
         if let key = key {
-            ret = crypto_generichash(output!.mutableBytesPtr, UInt(output!.length), message.bytesPtr, CUnsignedLongLong(message.length), key.bytesPtr, UInt(key.length))
+            ret = crypto_generichash(output!.mutableBytesPtr, output!.length, message.bytesPtr, CUnsignedLongLong(message.length), key.bytesPtr, key.length)
         } else {
-            ret = crypto_generichash(output!.mutableBytesPtr, UInt(output!.length), message.bytesPtr, CUnsignedLongLong(message.length), nil, 0)
+            ret = crypto_generichash(output!.mutableBytesPtr, output!.length, message.bytesPtr, CUnsignedLongLong(message.length), nil, 0)
         }
         if ret != 0 {
             return nil
@@ -65,9 +65,9 @@ public class GenericHash {
             }
             var ret: CInt
             if let key = key {
-                ret = crypto_generichash_init(state!, key.bytesPtr, UInt(key.length), UInt(outputLength))
+                ret = crypto_generichash_init(state!, key.bytesPtr, key.length, outputLength)
             } else {
-                ret = crypto_generichash_init(state!, nil, 0, UInt(outputLength))
+                ret = crypto_generichash_init(state!, nil, 0, outputLength)
             }
             if ret != 0 {
                 return nil
@@ -85,7 +85,7 @@ public class GenericHash {
     
         public func final() -> NSData? {
             let output = NSMutableData(length: outputLength)
-            if crypto_generichash_final(state!, output!.mutableBytesPtr, UInt(output!.length)) != 0 {
+            if crypto_generichash_final(state!, output!.mutableBytesPtr, output!.length) != 0 {
                 return nil
             }
             return output
